transpile to multiple files (WIP)
