+++ .vscode/launch.json
@@ -8,7 +8,7 @@
             "program": "${workspaceRoot}/bin/ts2c",
             //"program": "${workspaceRoot}/sputniktests/run_test",
             "stopOnEntry": false,
-            "args": ["tests/mem/mem7.ts"],
+            "args": ["tests/strings/string6.ts"],
             "cwd": "${workspaceRoot}",
             "sourceMaps": true
         }
