+++ tests/Makefile
@@ -6,8 +6,7 @@ endif
 
 TS := $(sort $(wildcard *.ts) $(wildcard */*.ts))
 TS := $(filter-out realworld/parseHtml.ts, $(TS)) # not passing yet
-TS := $(filter-out statements/import.ts, $(TS)) # not passing yet
-TS := $(filter-out statements/import_target.ts, $(TS)) # not passing yet
+TS := $(filter-out $(wildcard imports/*.ts), $(TS)) # can't handle multi-file tests via Makefile yet
 C = $(TS:.ts=.c)
 ifeq ($(target), Windows)
     ELF = $(TS:.ts=.exe)
