+++ tests/imports/common.h
@@ -1,0 +1,8 @@
+#ifndef TS2C_COMMON_H
+#define TS2C_COMMON_H
+
+#include <stdio.h>
+
+typedef short int16_t;
+
+#endif
