+++ tests/imports/files/simple-test.c
@@ -1,0 +1,11 @@
+#include "common.h";
+
+#include "simple-test.h"
+
+int16_t a;
+
+int main(void) {
+    a = 10;
+
+    return 0;
+}
