+++ tests/imports/files/simple-test.ts
@@ -1,0 +1,1 @@
+export var a = 10;
