+++ tests/imports/import_target.c
@@ -1,0 +1,10 @@
+#include <stdio.h>
+
+#include <stdlib.h>
+
+int main(void) {
+    printf(/* Unsupported printf expression */);
+    printf(/* Unsupported printf expression */);
+
+    return 0;
+}
