+++ tests/imports/simple.c
@@ -1,0 +1,9 @@
+#include "common.h";
+
+#include "files/simple-test.h"
+
+int main(void) {
+    printf("%d\n", a);
+
+    return 0;
+}
