+++ tests/imports/simple.ts
@@ -1,0 +1,3 @@
+import { a } from './files/simple-test';
+
+console.log(a);
