--- tests/statements/files/import-test.ts
@@ -1,1 +1,0 @@
-export var a = 10;
