--- tests/statements/import.c
@@ -1,12 +1,0 @@
-#include <stdio.h>
-
-typedef short int16_t;
-
-static int16_t a;
-
-int main(void) {
-    a = 10;
-    printf("%d\n", a);
-
-    return 0;
-}
