--- tests/statements/import.ts
@@ -1,3 +1,0 @@
-import { a } from './files/import-test';
-
-console.log(a);
