--- tests/statements/import_target.c
@@ -1,10 +1,0 @@
-#include <stdio.h>
-
-#include <stdlib.h>
-
-int main(void) {
-    printf("%d\n", abs(-5));
-    printf("%d\n", abs(10));
-
-    return 0;
-}
